World of Warcraft: Fix some bad buffer management
